iio: buffer-dma: Fix warning about integer to pointer cast
queue->poll_wakeup_flags is a unsigned int put gets cast to void * when
being passed on wake_up_interruptible_poll().

On a 64-bit system this results in the following warning:

drivers/iio/buffer/industrialio-buffer-dma.c: In function ‘iio_dma_buffer_block_done’:
./include/linux/wait.h:214:38: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  __wake_up(x, TASK_INTERRUPTIBLE, 1, (void *) (m))
                                      ^
drivers/iio/buffer/industrialio-buffer-dma.c:228:2: note: in expansion of macro ‘wake_up_interruptible_poll’
  wake_up_interruptible_poll(&queue->buffer.pollq, queue->poll_wakup_flags);
  ^~~~~~~~~~~~~~~~~~~~~~~~~~
  CC      drivers/iio/buffer/industrialio-buffer-dmaengine.o

Fix this by explicitly casting to uintptr_t which is a integer with the
same size as a pointer.

This change is purely cosmetic. The generated code is the same before and
after.

+++ drivers/iio/buffer/industrialio-buffer-dma.c
@@ -225,7 +225,7 @@ void iio_dma_buffer_block_done(struct iio_dma_buffer_block *block)
 	spin_unlock_irqrestore(&queue->list_lock, flags);
 
 	iio_buffer_block_put_atomic(block);
-	wake_up_interruptible_poll(&queue->buffer.pollq, queue->poll_wakup_flags);
+	wake_up_interruptible_poll(&queue->buffer.pollq, (uintptr_t)queue->poll_wakup_flags);
 }
 EXPORT_SYMBOL_GPL(iio_dma_buffer_block_done);
